Add ostream formatters for TargetPtr/TargetVal. (#5592)
